add streaming and platform components
